Fixed some paths issues for tomtom_minimal

+++ ci/build_tomtom_minimal.sh
@@ -10,39 +10,28 @@ set -e
 export ARCH=arm-linux
 
 mkdir -p build
-cd build
+pushd build
 cmake ../ -DCMAKE_INSTALL_PREFIX=$PREFIX -DFREETYPE_INCLUDE_DIRS=$PREFIX/include/freetype2/ -Dsupport/gettext_intl=TRUE \
 -DHAVE_API_TOMTOM=TRUE -DXSLTS=tomtom -DAVOID_FLOAT=TRUE -Dmap/mg=FALSE -DUSE_PLUGINS=0 -DCMAKE_TOOLCHAIN_FILE=../Toolchain/$ARCH.cmake \
 -DDISABLE_QT=ON -DSAMPLE_MAP=n -DBUILD_MAPTOOL=n
 make -j$JOBS
 make install
-cd ..
-
 
 # creating directories
 OUT_PATH="/tmp/tomtom/sdcard"
-rm -rf $OUT_PATH
-mkdir -p $OUT_PATH
-cd $OUT_PATH
-mkdir -p navit
-cd navit
-mkdir -p bin share
-cd share 
-mkdir -p fonts
-cd ..
-
+[ -d $OUT_PATH ] && rm -rf $OUT_PATH
+mkdir -p $OUT_PATH/navit/bin
+mkdir -p $OUT_PATH/navit/share/fonts
+mkdir -p $OUT_PATH/navit/share/icons
+mkdir -p $OUT_PATH/navit/share/maps
 
 # navit executable
-cp $PREFIX/bin/navit bin/
-
+cp navit/navit $OUT_PATH/navit/bin
 
 # fonts
-cp -r ~/navit/navit/fonts/*.ttf $OUT_PATH/navit/share/fonts
+cp -r ../navit/fonts/*.ttf $OUT_PATH/navit/share/fonts
 
 # images and xml
-cd share
-mkdir icons
-cd icons
 cp $PREFIX/share/navit/icons/*16.png ./
 cp $PREFIX/share/navit/icons/*32.png ./
 cp $PREFIX/share/navit/icons/*48.png ./
@@ -51,8 +40,6 @@ cp $PREFIX/share/navit/icons/nav*.* ./
 cp $PREFIX/share/navit/icons/country*.png ./
 cd ..
 cp $PREFIX/share/navit/navit.xml ./tomtom480.xml
-mkdir -p maps
-
 
 # locale
 cp -r $PREFIX/share/locale ./
